Move `Arena` & `ArenaList` to `Common`
