+++ ARMeilleure/IntermediateRepresentation/Operation.cs
@@ -1,3 +1,4 @@
+using ARMeilleure.Common;
 using System;
 using System.Runtime.CompilerServices;
